# lib/api-zod/src/generated/types/keyResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */

export interface KeyResult {
  id: number;
  okrId: number;
  descricao: string;
  meta: string;
  unidade: string;
  valor_atual: string;
  /** nao_iniciado | em_andamento | concluido | cancelado */
  status: string;
}