# lib/api-zod/src/generated/types/keyResultInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */

export interface KeyResultInput {
  /** @minLength 1 */
  descricao: string;
  meta: string;
  unidade: string;
  valor_atual: string;
  status: string;
}