# lib/api-zod/src/generated/types/keyResultUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */

export interface KeyResultUpdate {
  /** @minLength 1 */
  descricao?: string;
  meta?: string;
  unidade?: string;
  valor_atual?: string;
  status?: string;
}