# lib/api-zod/src/generated/types/kpiInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */

export interface KpiInput {
  /** @minLength 1 */
  nome: string;
  descricao?: string;
  categoria: string;
  meta: string;
  valor_atual: string;
  unidade: string;
  tendencia?: string;
  semaforo: string;
}