# lib/api-zod/src/generated/types/necessidadesStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */
import type { NecessidadesStatsPorEixo } from './necessidadesStatsPorEixo';
import type { NecessidadesStatsPorMoscow } from './necessidadesStatsPorMoscow';
import type { NecessidadesStatsPorStatus } from './necessidadesStatsPorStatus';

export interface NecessidadesStats {
  total: number;
  por_status: NecessidadesStatsPorStatus;
  por_moscow: NecessidadesStatsPorMoscow;
  por_eixo: NecessidadesStatsPorEixo;
  orcamento_total_planejado: number;
  orcamento_total_realizado: number;
}