# lib/api-zod/src/generated/types/okr.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */

export interface Okr {
  id: number;
  objetivo: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  trimestre?: string | null;
  ano: number;
  /** ativo | concluido | cancelado */
  status: string;
  /** 0-100 */
  progresso?: number;
  createdAt: Date;
  updatedAt: Date;
}