# lib/api-zod/src/generated/types/okrWithKeyResults.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for PDTIC/AEB MoSCoW Management System
 * OpenAPI spec version: 0.1.0
 */
import type { KeyResult } from './keyResult';

export interface OkrWithKeyResults {
  id: number;
  objetivo: string;
  /** @nullable */
  descricao?: string | null;
  /** @nullable */
  trimestre?: string | null;
  ano: number;
  status: string;
  progresso?: number;
  createdAt: Date;
  updatedAt: Date;
  keyResults: KeyResult[];
}